# packages/backend/openapi.yaml
openapi: 3.1.0
info:
  title: ProAppStore Platform API
  version: 1.0.0
  description: 'REST API for ProAppStore (PAS) — the Pro app marketplace platform.


    Powers identity, app provisioning + deploy, the per-app backend services (KV, counters, storage, rooms,
    secrets, AI, email/SMS, maps, push), the storefront + submissions review flow, the Agent Teams build
    loop, billing, and creator payouts.


    Auth: send `Authorization: Bearer <PAS session JWT>` (HS256, from `/v1/auth/exchange` or an OAuth
    flow). A handful of service-to-service routes use the `X-Internal-Token` header instead. Public routes
    need no auth.


    This spec is generated from the backend route table; request/response bodies are typed for the core
    resources and generic (object) elsewhere.'
servers:
- url: https://api.proappstore.online
  description: Production
security:
- bearerAuth: []
...
paths:
...
  /v1/apps/{appId}/scheduled-runs:
    get:
      tags:
      - Tools
      summary: List scheduled registered-action runs (#123)
      description: Owner only. Returns recent platform-scheduled action history. Optional status is due, claimed, succeeded or failed; limit defaults to 50 and is capped at 200.
      operationId: get_v1_apps_appId_scheduled_runs
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [runs]
                properties:
                  runs:
                    type: array
                    items:
                      type: object
                      properties:
                        run_id:
                          type: string
                        app_id:
                          type: string
                        action_name:
                          type: string
                        source:
                          type: string
                        due_at:
                          type: integer
                        claimed_at:
                          type: integer
                          nullable: true
                        finished_at:
                          type: integer
                          nullable: true
                        status:
                          type: string
                          enum: [due, claimed, succeeded, failed]
                        changes:
                          type: integer
                          nullable: true
                        error:
                          type: string
                          nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - bearerAuth: []
      parameters:
      - name: appId
        in: path
        required: true
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          enum: [due, claimed, succeeded, failed]
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 200
...
  /v1/internal/payouts/run:
    post:
      tags:
      - Payouts
      summary: Run payouts
      operationId: post_v1_internal_payouts_run
      responses:
        '200':
          description: Success
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - internalToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
components:
  securitySchemes:
    githubOidc:
      type: http
      scheme: bearer
      bearerFormat: GitHub Actions OIDC JWT
      description: 'A GitHub Actions OIDC token (permissions: id-token: write; audience https://api.proappstore.online). Accepted by the keyless deploy routes and POST /v1/auth/exchange/oidc.'
    appToken:
      type: http
      scheme: bearer
      bearerFormat: pas_at_…
      description: 'Personal app token (#154): per-user, per-app, revocable; accepted by POST /v1/apps/{appId}/actions/{name} only. Never by MCP or any other route.'
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: PAS session JWT (HS256). uid/login/roles/appRoles claims.
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: Shared secret for service-to-service calls.
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
      - error
...